refactor(sdk): Rename oidc module to oauth
Since we mostly use OAuth 2.0 now.
