+++ bindings/matrix-sdk-ffi/src/client_builder.rs
@@ -3,7 +3,7 @@ use std::{fs, num::NonZeroUsize, path::Path, sync::Arc, time::Duration};
 use async_compat::get_runtime_handle;
 use futures_util::StreamExt;
 use matrix_sdk::{
-    authentication::oidc::qrcode::{self, DeviceCodeErrorResponseType, LoginFailureReason},
+    authentication::oauth::qrcode::{self, DeviceCodeErrorResponseType, LoginFailureReason},
     crypto::{
         types::qr_login::{LoginQrCodeDecodeError, QrCodeModeData},
         CollectStrategy, TrustRequirement,
