+++ crates/matrix-sdk/src/authentication/oauth/qrcode/messages.rs
