+++ crates/matrix-sdk/src/authentication/oauth/qrcode/rendezvous_channel.rs
