+++ crates/matrix-sdk/src/authentication/oauth/qrcode/secure_channel.rs
