+++ crates/matrix-sdk/src/client/futures.rs
@@ -29,7 +29,7 @@ use tracing::trace;
 
 use super::super::Client;
 #[cfg(feature = "experimental-oidc")]
-use crate::authentication::oidc::OidcError;
+use crate::authentication::oauth::OidcError;
 use crate::{
     config::RequestConfig,
     error::{HttpError, HttpResult},
