+++ crates/matrix-sdk/src/error.rs
@@ -354,7 +354,7 @@ pub enum Error {
     /// An error occurred interacting with the OpenID Connect API.
     #[cfg(feature = "experimental-oidc")]
     #[error(transparent)]
-    Oidc(#[from] crate::authentication::oidc::OidcError),
+    Oidc(#[from] crate::authentication::oauth::OidcError),
 
     /// A concurrent request to a deduplicated request has failed.
     #[error("a concurrent request failed; see logs for details")]
@@ -561,7 +561,7 @@ pub enum RefreshTokenError {
     /// An error occurred interacting with the OpenID Connect API.
     #[cfg(feature = "experimental-oidc")]
     #[error(transparent)]
-    Oidc(#[from] Arc<crate::authentication::oidc::OidcError>),
+    Oidc(#[from] Arc<crate::authentication::oauth::OidcError>),
 }
 
 /// Errors that can occur when manipulating push notification settings.
